unify Icon_size and refactor regex to constants
