+++ src/constants/iconSize.ts
@@ -1,0 +1,5 @@
+export const ICON_SIZE = {
+    xs: "2rem",  // mobile
+    md: "2.2rem" // desktop and above
+  };
+
